Decode the input of split_html prior handing it to TagIter
Otherwise the results splitted by TagIter will be unreliable. For
example it returned

ent">\xc2

for me. Which is the end of a tag, e.g. "<foo bar="document"> and this
tag is followed by \xc2\xb6, i.e. the sole textual content of the
element, which is the UTF-8 representation of U+OOB6, but TagIter only
returned the first byte of this codepoint.